+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.232.0] - 2023-11-14
+- Changed: backported Utils/DatabaseUtil from v3 to v2
+- Deprecated: old DatabaseUtil class
+
 ## [2.231.0] - 2023-10-24
 - Added: Utils/ModelUtil::findParentsRecursively()
 - Deprecated: ModelUtil::findParentsRecursively()